10-ai-infrastructure-auditor/requirements.txt
streamlit>=1.32.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
